Updated some files and fixed issues with Clear Database
Fixed a major issue in Clear Database to where it does actually clear and redirect the admins back to the dashboard. Also updated Sanitize, Remove, and Edit equipment. All 3 do not work as expected yet. Sanitize and Remove show blank screens even though there is code that is supposed to be shown