Added base for student side
Added student base for system, also updated some files to reflect student base. I added ID verification within the system to verify students are able to use the equipment, but have not finished it yet. It should be done tomorrow after I refresh myself on how to deal with files that will be turned into arrays